requirements.txt
appnope==0.1.0
astroid==2.3.2
backcall==0.1.0
decorator==4.4.1
Django==2.2.6
django-environ==0.4.5
django-extensions==2.2.5
djangorestframework==3.10.3
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
django-rest-swagger==2.2.0
djangorestframework-gis==0.14.0
django-cors-headers==3.1.1
django-model-utils==3.2.0
Faker==2.0.3
xeger==0.3.5
factory-boy==2.12.0
uuid==1.30
django-allauth==0.39.1
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
python-dateutil==2.8.1
django-money==1.0
pre-commit==1.20.0
black==19.10b0
flake8==3.7.9
tblib==1.5.0
